pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "aeternum"
version = "0.0.5"
description = "Python-based CLI tool for virtual test environments"
authors = ["Chino Franco <chino.franco@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "aeternum" }
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
colorama = "^0.4.6"
requests = "^2.32.3"
pyyaml = "^6.0.1"
jinja2 = "^3.1.4"
tqdm = "^4.66.4"
pydantic = "^2.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.8.0"
black = "^24.4.2"
coverage = "^7.6.0"
poetry = "^1.8.3"
pytest-mock = "^3.14.0"

[tool.poetry.scripts]
aeternum = "aeternum.main:cli"
